requirements.txt
streamlit
google-genai
python-dotenv
openai-whisper
edge-tts
sounddevice
soundfile
scipy
numpy
ffmpeg-python
